# blocks-import-script: add --markdown flag and remove --csv-output flag