## Summary

This PR adds `--markdown-output` support to blocks import script.
This PR also removes `--csv-output` flag.

This option can be used like this
```
python scripts/block-import-stats.py
short-benchmark/20241208T172603_9207eaf/blocks-import-benchmark.csv
short-benchmark/20241222T012247_aba9b582/blocks-import-benchmark.csv 
--markdown-output README.md
```

The output is a nice markdown table :

## blocks-import-benchmark.csv vs blocks-import-benchmark.csv

| Block Range | BPS Baseline | BPS Contender | TPS Baseline | TPS Contender | Time Baseline | Time Contender | BPS Diff | TPS Diff | Time Diff |
|------------|--------------|---------------|--------------|---------------|---------------|----------------|----------|----------|-----------|
| (20000001, 20111112] | 25.62 | 24.98 | 4176.33 | 4075.37 | 1h9m29s | 1h11m34s | -2.63% | -2.63% | 2.84% |
| (20111112, 20222223] | 25.91 | 26.95 | 4237.79 | 4408.99 | 1h13m53s | 1h11m3s | 4.01% | 4.01% | -3.84% |
| (20222223, 20333334] | 26.60 | 27.69 | 4314.13 | 4490.81 | 1h7m0s | 1h4m26s | 4.05% | 4.05% | -3.87% |
| (20333334, 20444445] | 29.13 | 30.55 | 4586.04 | 4811.19 | 1h5m47s | 1h2m47s | 4.83% | 4.83% | -4.59% |
| (20444445, 20555556] | 29.12 | 29.79 | 4483.73 | 4588.42 | 1h0m59s | 59m38s | 2.29% | 2.29% | -2.22% |
| (20555556, 20666667] | 29.62 | 30.58 | 4472.00 | 4619.38 | 1h4m33s | 1h2m32s | 3.26% | 3.26% | -3.14% |
| (20666667, 20777778] | 30.50 | 31.44 | 4797.46 | 4947.33 | 58m15s | 56m32s | 3.08% | 3.08% | -2.97% |
| (20777778, 20888889] | 30.99 | 32.87 | 4969.59 | 5273.27 | 1h1m50s | 58m18s | 6.08% | 6.08% | -5.72% |
| (20888889, 21000001] | 30.81 | 32.98 | 5080.18 | 5438.98 | 1h2m16s | 58m16s | 7.00% | 7.00% | -6.53% |

## Summary

| Metric | Value |
|--------|-------|
| Total Blocks | 991808 |
| Baseline Time | 9h44m5s |
| Contender Time | 9h25m10s |
| Time Difference | -18m55s |
| Time Difference % | -3.24% |

## Legend

- BPS Diff: Blocks per second difference (+)
- TPS Diff: Transactions per second difference
- Time Diff: Time to process difference (-)

(+) = more is better, (-) = less is better