@yakimant 2025-01-06
"(20000001, 20111112]" - different brackets?