@siddarthkay 2025-01-06
> `(20000001, 20111112]` - different brackets?

In python : The brackets `(` and `]` in the output (20000001, 20111112] represent an "interval notation".

`(` means exclusive (does not include) the first number
`]` means inclusive (includes) the last number