@arnetheduck 2025-01-06
my preference would broadly be that we cut down on the output formats to as few as possible - ie either csv or markdown - there's no point maintaining a bunch of formats that don't get used.