@siddarthkay 2025-01-06
I like `csv` as an output format.
However judging by what you want as part of Readme on Github, the markdown table output seems like what we should keep as an output format.